test/cnc3018/gst_imager.py: make nice
